add: シェアのWidgetを適切な場所に作成しました
